refactor: extract dateUtils
