+++ src/components/configuration/ConfigBar.tsx
@@ -17,9 +17,10 @@ import MobileConfigUpdateBar from "@/components/configuration/MobileConfigUpdate
 import { useIntegrationUser } from "@/lib/hooks/useIntegrationUser";
 import { useUserConfig } from "@/lib/hooks/useUserConfig";
 import { IntegrationIdentifier } from "@/lib/integrations/active";
-import { classConfigRecurrentId, classRecurrentId, zeroIndexedWeekday } from "@/lib/integrations/common";
+import { classConfigRecurrentId, classRecurrentId } from "@/lib/integrations/common";
 import { arraysAreEqual } from "@/lib/utils/arrayUtils";
 import { hexColorHash } from "@/lib/utils/colorUtils";
+import { zeroIndexedWeekday } from "@/lib/utils/dateUtils";
 import { ClassConfig, IntegrationConfig, RezervoClass } from "@/types/rezervo";
 
 function ConfigBar({
