+++ src/components/modals/ClassInfo/ClassInfo.tsx
@@ -12,10 +12,11 @@ import ClassUsersAvatarGroup from "@/components/schedule/class/ClassUsersAvatarG
 import ConfirmationDialog from "@/components/utils/ConfirmationDialog";
 import { useUserConfig } from "@/lib/hooks/useUserConfig";
 import { useUserSessions } from "@/lib/hooks/useUserSessions";
-import { getCapitalizedWeekday, isClassInThePast } from "@/lib/integrations/common";
+import { isClassInThePast } from "@/lib/integrations/common";
 import { stringifyClassPopularity } from "@/lib/popularity";
 import { formatNameArray } from "@/lib/utils/arrayUtils";
 import { hexWithOpacityToRgb } from "@/lib/utils/colorUtils";
+import { getCapitalizedWeekday } from "@/lib/utils/dateUtils";
 import { ClassPopularity, RezervoClass, SessionStatus, StatusColors, UserNameWithIsSelf } from "@/types/rezervo";
 
 export default function ClassInfo({
