+++ src/components/schedule/DaySchedule.tsx
@@ -3,7 +3,8 @@ import { DateTime } from "luxon";
 import React from "react";
 
 import ClassCard from "@/components/schedule/class/ClassCard";
-import { classRecurrentId, getCapitalizedWeekday } from "@/lib/integrations/common";
+import { classRecurrentId } from "@/lib/integrations/common";
+import { getCapitalizedWeekday } from "@/lib/utils/dateUtils";
 import {
     AllConfigsIndex,
     ClassPopularity,
