+++ src/lib/integrations/common.ts
@@ -1,6 +1,5 @@
 import { DateTime } from "luxon";
 
-import { LOCALE, TIME_ZONE } from "@/lib/consts";
 import { createClassPopularityIndex } from "@/lib/popularity";
 import { serializeSchedule } from "@/lib/serializers";
 import {
@@ -13,24 +12,6 @@ import {
     RezervoSchedule,
     RezervoWeekSchedule,
 } from "@/types/rezervo";
-
-export const calculateMondayOffset = () => DateTime.now().setZone(TIME_ZONE).weekday - 1;
-
-export const getDateTime = (date: string): DateTime => DateTime.fromISO(date, { zone: TIME_ZONE, locale: LOCALE });
-
-export const zeroIndexedWeekday = (oneIndexedWeekday: number): number => (oneIndexedWeekday + 6) % 7;
-
-export const capitalizeFirstCharacter = (text: string) => {
-    return `${text[0]!.toUpperCase()}${text.slice(1)}`;
-};
-
-export const getCapitalizedWeekday = (date: DateTime): string => {
-    if (!date.isValid || date.weekdayLong === null) {
-        throw new Error("Invalid date");
-    }
-
-    return capitalizeFirstCharacter(date.weekdayLong);
-};
 
 export async function fetchIntegrationPageStaticProps<T>(
     integrationProfile: IntegrationProfile,
