+++ src/lib/integrations/sit/fetchers.ts
@@ -1,8 +1,8 @@
 import { DateTime } from "luxon";
 
 import { GROUP_BOOKING_URL, TIME_ZONE } from "@/lib/consts";
-import { calculateMondayOffset } from "@/lib/integrations/common";
 import { SitDaySchedule, SitWeekSchedule } from "@/lib/integrations/sit/types";
+import { calculateMondayOffset } from "@/lib/utils/dateUtils";
 
 function sitScheduleUrl(token: string, fromISO: string | null = null) {
     return (
